refactor(problem): move `can_view_problem` to problem class
