+++ model/submission.py
@@ -14,7 +14,6 @@ from datetime import datetime, timedelta
 from mongo import *
 from mongo import engine
 from mongo.utils import (
-    can_view_problem,
     RedisCache,
     perm,
     drop_none,
@@ -54,7 +53,7 @@ def create_submission(user, language_type, problem_id):
     if not problem:
         return HTTPError('Unexisted problem id.', 404)
     # problem permissoion
-    if not can_view_problem(user, problem):
+    if not problem.check_view_permission(user=user):
         return HTTPError('problem permission denied!', 403)
     # check deadline
     for homework in problem.obj.homeworks:
