+++ mongo/utils.py
@@ -13,7 +13,6 @@ if TYPE_CHECKING:
 __all__ = (
     'hash_id',
     'perm',
-    'can_view_problem',
     'RedisCache',
     'doc_required',
 )
@@ -32,25 +31,6 @@ def perm(course, user):
         user.role == 0, user == course.teacher, user in course.tas,
         user.username in course.student_nicknames.keys(), True
     ].index(True)
-
-
-def can_view_problem(user: 'User', problem: 'Problem'):
-    '''cheeck if a user can view the problem'''
-    from .user import User
-    if user.role == User.engine.Role.ADMIN:
-        return True
-    if user.contest:
-        if user.contest in problem.contests:
-            return True
-        return False
-    if user.username == problem.owner:
-        return True
-    for course in problem.courses:
-        permission = 1 if course.course_name == 'Public' else perm(
-            course, user)
-        if permission and (problem.problem_status == 0 or permission >= 2):
-            return True
-    return False
 
 
 class Cache(abc.ABC):
